PersonalizedATSSystem_GeminiPro/requirements.txt
streamlit
google.generativeai
PyPDF2
python-dotenv
docx